Ransomware/requirements.txt
Pillow==12.2.0
